Move device_type_map to separate .py file, because of circular dependency between EFirmwareId and HausBusUtils
